Split lower.rs into items and types sub-modules
Reduces lower.rs from 2095 to 1161 lines (647 code + 514 tests)
by extracting:

- lower/items.rs (546 lines): item, import, const, fn, type decl,
  for block, trait, test block lowering
- lower/types.rs (396 lines): type expression and type def lowering

Joins existing lower/expr.rs, lower/jsx.rs, lower/pattern.rs modules.
Purely structural - no behavioral changes.
